+++ src/config.js
@@ -1,3 +1,4 @@
+// defined production and development variables for re-checking the documents page. These can be customized as needed.
 const prod = {
   url: {
     DELAY: 60 * 30
@@ -6,7 +7,7 @@ const prod = {
 
 const dev = {
   url: {
-    DELAY: 10
+    DELAY: 60 * 30
   }
 };
